Removed `EclipseProjectIdentifier`
For the Eclipse model, we assume that the name of an EclipseProject is
sufficient for identification, and that the path of a ProjectDependency
is sufficient to locate the target project.

Based on that assumption, this commit removes the separate identifier
that was added to the eclipse model. It also simplifies the code for
wiring up dependencies in the Eclipse model, removing the need to track
the ProjectComponentIdentifier for each project dependency.